+++ ib1/ib1.psm1
@@ -12,13 +12,40 @@ $mslearnGit='MicrosoftLearning'
 $defaultSwitchId='c08cb7b8-9b3c-408e-8e30-5e16a3aeb444'
 $logStart=$true
 $repoParam=@{
-  'M10979'=@{'repo'='10979-MicrosoftAzureFundamentals';'srcPath'='AllFiles';'destPath'='E:\';'shortcut'='MicrosoftLearning/10979-MicrosoftAzureFundamentals/tree/master/Instructions';'vmName'='mia-cl1'}
-  'm20533'=@{'repo'='20533-ImplementingMicrosoftAzureInfrastructureSolutions';'srcPath'='Allfiles';'destPath'='F:\';'shortcut'='MicrosoftLearning/20533-ImplementingMicrosoftAzureInfrastructureSolutions/tree/master/Instructions';'vmName'='mia-cl1'}}
-$gitParam=@{
-  'm10979'='new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/10979-MicrosoftAzureFundamentals/tree/master/Instructions" -title "Ateliers stage m10979";if ($env:COMPUTERNAME -like "mia-cl1") {get-ib1Git 10979-MicrosoftAzureFundamentals -srcPath Allfiles -destPath E:\}';
-  'm20533'='new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/20533-ImplementingMicrosoftAzureInfrastructureSolutions/tree/master/Instructions" -title "Ateliers stage m20533";if ($env:COMPUTERNAME -like "mia-cl1") {get-ib1Git 20533-ImplementingMicrosoftAzureInfrastructureSolutions -srcPath Allfile -destPath F:\}';
-  'msaz100'='get-ib1Git AZ-100-MicrosoftAzureInfrastructureDeployment;$dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-100-MicrosoftAzureInfrastructureDeployment\";remove-item "$destAZ-100T03A-ENU-LabFiles.zip -force;remove-item "$destAZ-100T04A-ENU-LabFiles.zip -force';
-  'msaz101'='get-ib1Git AZ-101-MicrosoftAzureIntegrationandSecurity;$dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-101-MicrosoftAzureIntegrationandSecurity\";remove-item "$destAZ-101T03A-ENU-LabFiles.zip" -force;remove-item "$destAZ-101T04A-ENU-LabFiles.zip" -force'}
+  'm10979'='
+  new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/10979-MicrosoftAzureFundamentals/tree/master/Instructions" -title "Ateliers stage m10979";
+  if ($env:COMPUTERNAME -like "pc-formateur") {get-ib1Repo 10979-MicrosoftAzureFundamentals -srcPath Allfiles -destPath E:\}';
+  'm20533'='
+  new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/20533-ImplementingMicrosoftAzureInfrastructureSolutions/tree/master/Instructions" -title "Ateliers stage m20533";
+  if ($env:COMPUTERNAME -like "pc-formateur") {get-ib1Repo 20533-ImplementingMicrosoftAzureInfrastructureSolutions -srcPath Allfiles -destPath F:\}';
+  'msaz100'='
+  get-ib1Repo AZ-100-MicrosoftAzureInfrastructureDeployment;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-100-MicrosoftAzureInfrastructureDeployment\";
+  rename-item -path $dest -newName "Ateliers MSAZ100" -errorAction SilentlyContinue;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\Ateliers MSAZ100\";
+  remove-item "$($dest)AZ-100T03A-ENU-LabFiles.zip" -force -errorAction SilentlyContinue;
+  remove-item "$($dest)AZ-100T04A-ENU-LabFiles.zip" -force -errorAction SilentlyContinue;
+  Add-Type -AssemblyName System.IO.Compression.FileSystem;
+  get-childitem ($dest)|foreach-object {unzip $_.fullName $dest;remove-item $_.fullName -force -errorAction SilentlyContinue}
+  get-childitem ($dest) -directory|foreach-object {move-item "$($_.fullname)\*" -destination $dest;remove-item $_.fullName -force}
+  get-childitem ($dest) -file|foreach-object {rename-item -path $_.fullName -newName "Partie $($_.name[8]).pdf"}';
+  'msaz101'='
+  get-ib1Repo AZ-101-MicrosoftAzureIntegrationandSecurity;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-101-MicrosoftAzureIntegrationandSecurity\";
+  rename-item -path $dest -newName "Ateliers MSAZ101" -errorAction SilentlyContinue;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\Ateliers MSAZ101\";
+  remove-item "$($dest)AZ-101T03A-ENU-LabFiles.zip" -force;
+  remove-item "$($dest)AZ-101T04A-ENU-LabFiles.zip" -force;
+  Add-Type -AssemblyName System.IO.Compression.FileSystem
+  get-childitem ($dest)|foreach-object {unzip $_.fullName $dest;remove-item $_.fullName -force -errorAction SilentlyContinue}
+  get-childitem ($dest) -directory|foreach-object {move-item "$($_.fullname)\*" -destination $dest;remove-item $_.fullName -force}
+  get-childitem ($dest) -file|foreach-object {rename-item -path $_.fullName -newName "Partie $($_.name[8]).pdf"}'}
+
+
+function Unzip {
+param([string]$zipfile,[string]$outpath)
+  [System.IO.Compression.ZipFile]::ExtractToDirectory($zipfile, $outpath)
+}
 
 function write-ib1log {
 [CmdletBinding(DefaultParameterSetName='TextLog')]
@@ -181,7 +208,7 @@ param([string]$VMName,[bool]$exactVMName=$false)
     write-ib1log "Impossible de trouver une machine virtuelle '$VMName'." -ErrorLog}}
   return $gResult}
 
-function get-ib1Git {
+function get-ib1Repo {
 <#
 .SYNOPSIS
 Cette commande permet de copier en local le contenu d'un repositery Git.
@@ -192,13 +219,14 @@ Chemin des fichiers à récupérer dans le Repositery (par défaut "/")
 .PARAMETER destPath
 Chemin local ou seront posés les fichier récupérés du Git (par défaut, sur le bureau de l'utilisateur actuel)
 .PARAMETER course
-Référence du stage pour paramètrage simplifié (remplaçe les 3 paramètres précédents pour les environnements prévus).
+Référence du stage pour paramètrage simplifié (remplaçe les 3 paramètres précédents).
+Valeurs possibles : m10979, m20533, msaz100 et msAZ101
 .EXAMPLE
-get-ib1Git -repo '10979-MicrosoftAzureFundamentals' -srcPath 'Allfiles' -destPath 'c:\10979'
+get-ib1Repo -repo '10979-MicrosoftAzureFundamentals' -srcPath 'Allfiles' -destPath 'c:\10979'
 Récupère les fichiers contenu dans le répertoire 'AllFiles' du repo '10979' et le copie dans le repertoire local c:\10979
 .EXAMPLE
-get-ib1Git -course m10979
-Récupère les fichiers pour le stage m10979
+get-ib1Repo -course msAZ100
+Récupère les fichiers pour le stage msAZ100
 #>
 [CmdletBinding(
 DefaultParameterSetName='Repo')]
@@ -210,16 +238,17 @@ PARAM(
 begin{get-ib1elevated $true; compare-ib1PSVersion "4.0"}
 process {
 if ($course -ne '') {
-  if (-not $gitParam.$course) {write-ib1log "Le paramètre simplifié -course ne peut avoir que l'une des valeurs suivantes: $($gitParam.Keys). Merci de vérifier!" -ErrorLog}
+  if (-not $repoParam.$course) {write-ib1log "Le paramètre simplifié -course ne peut avoir que l'une des valeurs suivantes: $($repoParam.Keys). Merci de vérifier!" -ErrorLog}
   else {
     write-ib1log "Mise en place de l'environnement de stage pour le stage '$course'."
-    Invoke-Expression $gitParam.$course
+    Invoke-Expression $repoParam.$course
     break}}
 if (-not $Repo -or $Repo -eq '') {write-ib1log "Le paramétre -Repo est manquant, merci de vérifier !" -ErrorLog}
 if (-not $Repo.Contains('/')) {$repo=$mslearnGit+'/'+$Repo}
 #création du nom du répertoire local si non spécifié
 if ($destPath -eq '') {$destPath=[Environment]::GetFolderPath("CommonDesktopDirectory")+'\'+$Repo.substring($Repo.IndexOf('/')+1)}
 $ErrorActionPreference='silentlyContinue'
+[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
 $gitRequest=[System.Net.WebRequest]::Create("https://github.com/$Repo")
 $gitResponse=$gitRequest.GetResponse()
 if ([int]$gitResponse.StatusCode -ne 200) {write-ib1log "L'url 'https://github.com/$Repo' n'existe pas ou est injoignable." -ErrorLog}
@@ -229,7 +258,6 @@ if (-not (Test-Path $destPath)) {
   else {
     write-ib1log "Le répertoire '$destPath' n'existe pas et sera créé." -DebugLog
     New-Item $destPath -ItemType Directory|Out-Null}}
-[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
 $srcFolder="https://api.github.com/repos/$Repo/zipball"
 $destZip=$env:TEMP+'\'+$Repo.substring($Repo.IndexOf('/')+1)+'.zip'
 write-ib1log "Récupération du fichier '$srcFolder' dans '$destZip'." -DebugLog
@@ -238,80 +266,17 @@ Add-Type -AssemblyName System.IO.Compression.FileSystem
 if ($srcPath -eq '*') {$srcFind='*'} else {$srcFind="*/$srcPath*"}
 $zip=[IO.Compression.ZipFile]::OpenRead($destZip)
 $zip.Entries|Where-Object {$_.FullName -like $srcFind}|ForEach-Object {
-  echo $_.fullname
   if ($srcPath -ne '*') {$destName=$_.FullName -replace "(?:[^\/]*)\/$srcPath"} else {$destName=$_.FullName.substring($_.FullName.IndexOf('/'))}
   if ($_.Name -eq '') {
     write-ib1log -progressTitleLog "Décompression des fichiers." "Création du dossier '$destName'."
     New-Item -Path $destPath$destName -ItemType Directory -Force -ErrorAction SilentlyContinue |Out-Null}
   elseif ($_.Name -notlike '*readme.md') {
     write-ib1log -progressTitleLog "Décompression des fichiers." "Décompression du fichier '$($_.Name)'."
-    if(![System.IO.File]::Exists($destPath+$destName)) {[IO.Compression.ZipFileExtensions]::ExtractToFile($_,$destPath+$destName)}}}
+    if(![System.IO.File]::Exists($destPath+$destName)) {[IO.Compression.ZipFileExtensions]::ExtractToFile($_,$destPath+$destName)|Out-Null}}}
 $zip.dispose()
 write-ib1log -progressTitleLog "Décompression des fichiers."
 Remove-Item -Path $destZip -Force}}
 
-
-function get-ib1Repo {
-<#
-.SYNOPSIS
-Cette commande permet de récupérer en local le contenu d'un repositery Git.
-.PARAMETER Repo
-Nom du repositery sur GitHub
-.PARAMETER srcPath
-Chemin des fichiers à récupérer dans le Repositery
-.PARAMETER destPath
-Chemin local ou seront posés les fichier récupérés du Git.
-.PARAMETER course
-Référence du stage pour paramètrage simplifié (remplaçe les 3 paramètres précédents).
-Nota : Si utilisé dans la VM, le paramètre "course" récupère les fichiers des labs,
-       Si utilisé hors VM, le paramètre "course" crée un raccourci vers les instructions des ateliers.
-.EXAMPLE
-get-ib1Repo -repo '10979-MicrosoftAzureFundamentals' -srcPath 'Allfiles' -destPath 'c:\10979'
-Récupère les fichiers contenu dans le répertoire 'AllFiles' du repo '10979' et le copie dans le repertoire local c:\10979
-.EXAMPLE
-get-ib1Repo -course m10979
-Récupère les fichiers pour le stage m10979
-#>
-[CmdletBinding(
-DefaultParameterSetName='VMName')]
-PARAM(
-[string]$Repo,
-[string]$srcPath,
-[string]$destPath='E:',
-[string]$course='')
-begin{get-ib1elevated $true; compare-ib1PSVersion "4.0"}
-process {
-if ($course -ne '') {
-  if (-not $repoParam.$course) {write-ib1log "Le paramètre simplifié -course ne peut avoir que l'une des valeurs suivantes: $($repoParam.Keys). Merci de vérifier!" -ErrorLog}
-  if ($env:COMPUTERNAME -notlike $repoParam.$course.vmName) {
-    write-ib1log "Création du raccourci pour les instructions d'atelier '$course' sur le bureau."
-    new-ib1Shortcut -URL "https://github.com/$($repoParam.$course.shortcut)" -title "instructions d'atelier - $course"
-    break}
-  $repo=$repoParam.$course.repo
-  $srcPath=$repoParam.$course.srcPath
-  $destpath=$repoParam.$course.destPath}
-if (-not $repo -or $repo -eq '') {write-ib1log "Le paramétre -Repo est manquant, merci de vérifier !" -ErrorLog}
-if (-not $srcPath -or $srcPath -eq '') {write-ib1log "Le paramétre -srcPath est manquant, merci de vérifier !" -ErrorLog}
-[Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-$srcFolder=$mslearnGitUrl+$repo+'/zipball'
-$destZip=$env:TEMP+'\'+$repo+'.zip'
-write-ib1log "Récupération du fichier '$srcFolder' dans '$destZip'." -DebugLog
-Invoke-WebRequest -Uri $srcFolder -OutFile $destZip
-Add-Type -AssemblyName System.IO.Compression.FileSystem
-write-ib1log "Création du dossier '$destPath'." -DebugLog
-if ($destPath.Length -gt 3) {New-Item -Path $destPath -ItemType Directory -Force|Out-Null}
-$zip=[IO.Compression.ZipFile]::OpenRead($destZip)
-$zip.Entries|Where-Object {$_.FullName -like "*$repo*/$srcPath/*"}|ForEach-Object {
-  $destName=$_.FullName -replace "(?:[^\/]*)\/$srcPath"
-  if ($_.Name -eq '') {
-    write-ib1log -progressTitleLog "Décompression des fichiers." "Création du dossier '$destName'."
-    New-Item -Path $destPath$destName -ItemType Directory -Force -ErrorAction SilentlyContinue |Out-Null}
-  else {
-    write-ib1log -progressTitleLog "Décompression des fichiers." "Décompression du fichier '$($_.Name)'."
-    if(![System.IO.File]::Exists($destPath+$destName)) {[IO.Compression.ZipFileExtensions]::ExtractToFile($_,$destPath+$destName)}}}
-$zip.dispose()
-write-ib1log -progressTitleLog "Décompression des fichiers."
-Remove-Item -Path $destZip -Force}}
 
 function set-ib1VMCheckpointType {
 <#
@@ -1359,5 +1324,6 @@ else {
 Set-Alias ibReset reset-ib1VM
 Set-Alias set-ib1VhdBoot mount-ib1VhdBoot
 Set-Alias complete-ib1Setup complete-ib1Install
-Export-moduleMember -Function install-ib1Chrome,complete-ib1Install,invoke-ib1NetCommand,new-ib1Shortcut,Reset-ib1VM,Mount-ib1VhdBoot,Remove-ib1VhdBoot,Switch-ib1VMFr,Test-ib1VMNet,Connect-ib1VMNet,Set-ib1TSSecondScreen,Import-ib1TrustedCertificate, Set-ib1VMCheckpointType, Copy-ib1VM, repair-ib1VMNetwork, start-ib1SavedVMs, get-ib1log, get-ib1version, stop-ib1ClassRoom, new-ib1Nat, invoke-ib1Clean, invoke-ib1Rearm, get-ib1Repo, set-ib1VMExternalMac, get-ib1Git
+Set-Alias get-ib1Git get-ib1repo
+Export-moduleMember -Function install-ib1Chrome,complete-ib1Install,invoke-ib1NetCommand,new-ib1Shortcut,Reset-ib1VM,Mount-ib1VhdBoot,Remove-ib1VhdBoot,Switch-ib1VMFr,Test-ib1VMNet,Connect-ib1VMNet,Set-ib1TSSecondScreen,Import-ib1TrustedCertificate, Set-ib1VMCheckpointType, Copy-ib1VM, repair-ib1VMNetwork, start-ib1SavedVMs, get-ib1log, get-ib1version, stop-ib1ClassRoom, new-ib1Nat, invoke-ib1Clean, invoke-ib1Rearm, get-ib1Repo, set-ib1VMExternalMac
 Export-ModuleMember -Alias set-ib1VhdBoot,ibreset,complete-ib1Setup