+++ ib1/changeLog.md
@@ -1,5 +1,5 @@
 ﻿**Version 1.10.1:**
- - Ajout d'une commande get-ib1Git plus générique que get-ib1Repo pour copier le contenu des repo Git en local
+ - refonte de la commande get-ib1Repo pour gérer diverses sortes d'actions lors de préparation des stages, prête pour les stages m20533, m10979, msAZ100 et msAZ101
  - Suppression de la création du réseau virtuel "ib1Nat" dans l'installation du master
 
 **Version 1.9.1:**